+++ .github/CHANGELOG.md
@@ -11,10 +11,15 @@ deploy webhooks to print changelog successfully
 That's all, matcher will stop when detects next line started with `###` match
 
 ### 5.1.0 Release (22.05.2025)
+* TODO MAKE LANGUAGE:YML UPDATE BEFORE RELEASE!
+* Added Heads from HeadDatabase (selected on each server start 200 per category)
+* Added Heads Search function
 * Added paginated Heads menu to support more than one site of heads
 * Added onVoteInventoryInteractEvent to block movement of items while plot voting stage
 * Added a lot of new default themes at themes.yml (+140)
 * Added possibility to add same meaning words for gtb themes by using ", " in themes.yml
+* Added /bba settheme <arena> <theme> for Console use
+* Fixed Floor NPC spawns on any actions
 * Fixed NPE if OptionsMenu Particles are opened as none builder whileas in gtb
 * Fixed on first time setup without restart the NPE getArenaInGameState is null
 * Fixed ItemFrames and Paintings were not able to be removed by player on plot